+++ src/main/java/cn/nukkit/inventory/PlayerEnderChestInventory.java
@@ -28,8 +28,8 @@ public class PlayerEnderChestInventory extends BaseInventory {
         }
         super.onOpen(who);
         ContainerOpenPacket containerOpenPacket = new ContainerOpenPacket();
-        containerOpenPacket.windowId = who.getWindowId(this);
-        containerOpenPacket.type = this.getType().getNetworkType();
+        containerOpenPacket.windowId = (byte) who.getWindowId(this);
+        containerOpenPacket.type = (byte) this.getType().getNetworkType();
         BlockEnderChest chest = who.getViewingEnderChest();
         if (chest != null) {
             containerOpenPacket.x = (int) chest.getX();
@@ -59,7 +59,7 @@ public class PlayerEnderChestInventory extends BaseInventory {
     @Override
     public void onClose(Player who) {
         ContainerClosePacket containerClosePacket = new ContainerClosePacket();
-        containerClosePacket.windowId = who.getWindowId(this);
+        containerClosePacket.windowId = (byte) who.getWindowId(this);
         containerClosePacket.wasServerInitiated = who.getClosingWindowId() != containerClosePacket.windowId;
         who.dataPacket(containerClosePacket);
         super.onClose(who);
